Added Auto size documentation (#1660)
### Requirements

Please take note of our contributing guidelines: https://laravel-excel-docs.dev/docs/3.0/getting-started/contributing
Filling out the template is required. Any pull request that does not include enough information to be reviewed in a timely manner may be closed at the maintainers' discretion.

Mark the following tasks as done:

* [x] Checked the codebase to ensure that your feature doesn't already exist.
* [x] Checked the pull requests to ensure that another person hasn't already submitted the feature or fix.
* [x] Adjusted the Documentation.
* [ ] Added tests to ensure against regression.

### Description of the Change

<!--

We must be able to understand the design of your change from this description. 
If we can't get a good idea of what the code will be doing from the description here, 
the pull request may be closed at the maintainers' discretion. 
Keep in mind that the maintainer reviewing this PR may not be familiar with or have 
worked with the code here recently, so please walk us through the concepts.

-->

### Why Should This Be Added?

Feature was not yet documented for v3.0
<!-- Explain why this functionality should be added in Laravel-Excel -->

### Benefits

Improved documentation and ease of use
<!-- What benefits will be realized by the code change? -->

### Possible Drawbacks

None, no code was changed
<!-- What are the possible side-effects or negative impacts (e.g. breaking changes) of the code change? -->

### Verification Process
None, just are changes made in .md files. Should not affect anything.

<!--

What process did you follow to verify that your change has the desired effects?

- How did you verify that all new functionality works as expected?
- How did you verify that all changed functionality works as expected?
- How did you verify that the change has not introduced any regressions?

-->

### Applicable Issues
N/A
<!-- Enter any applicable Issues here -->